git-svn-id: svn+ssh://10.1.90.223/SVNroot/astra/server/stable@153426 e9a6705e-e829-0410-83f2-cebdff2db7e5

+++ src/qrys.h
@@ -77,7 +77,7 @@ class QParams: public std::list<QParam> {
 
 struct TQry {
     TQuery Qry;
-    u_int64_t count;
+    size_t count;
     TQry(): Qry(&OraSession), count(0) {};
 };
